+++ include/internal/searchers/annealing.h
@@ -31,7 +31,7 @@
 #include <vector>
 #include <random>
 
-#include "cltune/searcher.h"
+#include "internal/searcher.h"
 
 namespace cltune {
 // =================================================================================================
