+++ src/memory.cc
@@ -25,7 +25,8 @@
 //
 // =================================================================================================
 
-#include "cltune/memory.h"
+// The corresponding header file
+#include "internal/memory.h"
 
 namespace cltune {
 // =================================================================================================
