+++ src/opencl.cc
@@ -25,7 +25,8 @@
 //
 // =================================================================================================
 
-#include "cltune/opencl.h"
+// The corresponding header file
+#include "internal/opencl.h"
 
 namespace cltune {
 // =================================================================================================
